user Controller done

+++ app/Http/Controllers/Controller.php
@@ -6,7 +6,9 @@ use App\Models\User;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Foundation\Validation\ValidatesRequests;
+use Illuminate\Http\Request;
 use Illuminate\Routing\Controller as BaseController;
+use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 use Illuminate\Support\Facades\Redirect;
 
@@ -22,16 +24,29 @@ class Controller extends BaseController
         ]);
     }
 
-    public function create()
-    {
-        return Inertia::render('User/Add');
-    }
-
     public function edit($user)
     {
         return Inertia::render('User/Edit', [
             'user' => User::findOrFail($user)->first()
         ]);
+    }
+
+    public function update(Request $request, $user){
+        $user = User::findOrFail($user);
+        $request->validate([
+            'name' => 'required|unique:users,name|string',
+            'email' => 'required|unique:users,email|email',
+            'password' => 'required|confirmed',
+        ]);
+
+        $user->update([
+            'name' => $request->name,
+            'email' => $request->email,
+            'password' => $request->password,
+            'isAdmin' => $request->isAdmin,
+        ]);
+
+        return Redirect::route('/user')->with('success', 'تمت الاضافة بنجاح');
     }
 
 
@@ -44,7 +59,11 @@ class Controller extends BaseController
     public function delete($id)
     {
         $user = User::findOrFail($id);
-        $user->delete();
-        return Redirect::route('dashboard')->with('success', 'تم الحذف بنجاح');
+        if($user == Auth::user()){
+            return Redirect::back()->with('success', 'الحساب مستخدم حالياً');
+        }else{
+            $user->delete();
+            return Redirect::back()->with('success', 'تم الحذف بنجاح');
+        }
     }
 }
